Disallow explicit "notify yes;" for mirror zones
Previous way of handling notification settings for mirror zones was a
bit tricky: any value of the "notify" option was accepted, but it was
subsequently overridden with dns_notifytype_explicit.  Given the way
zone configuration is performed, this resulted in the following
behavior:

  - if "notify yes;" was set explicitly at any configuration level or
    inherited from default configuration, it was silently changed and so
    only hosts specified in "also-notify", if any, were notified,

  - if "notify no;" was set at any configuration level, it was
    effectively honored since even though zone->notifytype was silently
    set to dns_notifytype_explicit, the "also-notify" option was never
    processed due to "notify no;" being set.

Effectively, this only allowed the hosts specified in "also-notify" to
be notified, when either "notify yes;" or "notify explicit;" was
explicitly set or inherited from default configuration.

Make the user more aware of how the requested configuration plays with
mirror zones and clarify the way "notify" option handling is implemented
for mirror zones by:

  - reporting a configuration error when a "notify yes;" setting
    explicitly specified in the configuration is to be applied to a
    mirror zone,

  - silently changing the default, implicit "notify yes;" setting to
    "notify explicit;" for mirror zones, to prevent reporting
    configuration errors when the user has not set "notify" explicitly.